feat: nameless whens (scopes)
Rework scopes to work without being named. The last one is the one that
toggles the block on and off. Eg, before you would have done:
```
Text
text text
a when props.a
text a
b when props.b
text b
when props.c
```

Now, you'd do:
```
Text
text text
when props.a
text a
when props.b
text b
when props.c
```

Also, to toggle a block on and off you need to set the `when` key as the
last prop. If it is before the last one, it will be considered a
conditional scope instead.