refactor: last when to turn block off for onWhen
Before this commit:
```
Vertical
backgroundColor red
when props.isBlue
backgroundColor blue
when props.magicPropThatTurnsItOn
```
Would turn the block off when `props.magicPropThatTurnsItOn` is true.

After this commit, `when props.magicPropThatTurnsItOn` has to be `onWhen props.magicPropThatTurnsItOn` and
the value has to be on the base scope of the block, like:
```
Vertical
backgroundColor red
onWhen props.magicPropThatTurnsItOn
when props.isBlue
backgroundColor blue
```

The reason for this change is to distinguish between whens that tell
scoped props within a block and `onWhen` which turns the whole block
off when true.