feat: bye bye proximity nesting, we still love you
Proximity nesting was a concept we tried with Tom when we first drafted
Views to solve the problem that nesting blocks inside of other blocks
introduced to newcomers.

The concept is pretty simple: a new line separates blocks apart.
We see the UI as a stack of blocks and thought it would be somewhat
natural to push blocks in and out of containers.

However, over time we found that it confused people more than it helped.

Furthermore, Views Tools now indents its output for greater readability.

So we've decided to drop proximity nesting and keep the indentation to
represent that nesting.

As part of this, we've lost one (non-documented and experimental)
feature: text interpolation. However, it was due a refactor with support
for blocks other than text, so we'll bring it back in a different shape
at some stage.

You can use this codemod to convert all your code automatically to the
new indentation model
https://github.com/viewstools/codemods/blob/master/codemod-indent.js.