fix: ViewsData to catch errors in onChange event listener
This fix catches uncaught errors within the event handlers onChange and
onSubmit and throw them back in the render method so that an error boundary
can catch them. Otherwise they get lost in the scope of those handlers.

It also changes the calculation of the return value of useData which now
only changes if the value on the path being searched changes.
This may affect the onChange handler when used with formatOut if the
formatter is expecting the latest data to do something with it.
In general it shouldn't be a problem since you're meant to catch the
path you need anyway and not anything more specific to express the
intent of the data key.

We may come back to revisit this choice though.